refactor package structure for CSV data handling components
