+++ src/main/kotlin/data/csv_data/csv_parser/CsvParser.kt
@@ -1,0 +1,8 @@
+package data.csv_data.csv_parser
+
+typealias CsvData = String
+
+interface CsvParser<T> {
+    fun parse(data: List<CsvData>): List<T>
+    fun serialize(data: List<T>): List<CsvData>
+}
