+++ src/main/kotlin/data/csv_data/datasource/AuditLogDataSource.kt
@@ -1,0 +1,20 @@
+package data.csv_data.datasource
+
+import data.csv_data.csv_parser.AuditLogCsvParser
+import data.csv_data.csv_parser.CsvHandler
+import data.csv_data.dto.AuditLogDto
+
+class AuditLogDataSource(
+    private val csvHandler: CsvHandler,
+    private val csvParser: AuditLogCsvParser,
+) : DataSource<AuditLogDto> {
+    override fun fetch(): List<AuditLogDto> {
+        val lines = csvHandler.getLines()
+        return csvParser.parse(lines)
+    }
+
+    override fun save(data: List<AuditLogDto>) {
+        val serializedData = csvParser.serialize(data)
+        csvHandler.write(serializedData)
+    }
+}
