+++ src/main/kotlin/data/csv_data/datasource/DataSource.kt
@@ -1,0 +1,6 @@
+package data.csv_data.datasource
+
+interface DataSource<T> {
+    fun fetch(): List<T>
+    fun save(data: List<T>)
+}
