+++ src/main/kotlin/data/csv_data/datasource/TaskDataSource.kt
@@ -1,0 +1,20 @@
+package data.csv_data.datasource
+
+import data.csv_data.csv_parser.CsvHandler
+import data.csv_data.csv_parser.TaskCsvParser
+import data.csv_data.dto.TaskDto
+
+class TaskDataSource(
+    private val csvHandler: CsvHandler,
+    private val csvParser: TaskCsvParser,
+) : DataSource<TaskDto> {
+    override fun fetch(): List<TaskDto> {
+        val lines = csvHandler.getLines()
+        return csvParser.parse(lines)
+    }
+
+    override fun save(data: List<TaskDto>) {
+        val serializedData = csvParser.serialize(data)
+        csvHandler.write(serializedData)
+    }
+}
