+++ src/main/kotlin/data/csv_data/dto/AuditLogDto.kt
@@ -1,0 +1,12 @@
+package data.csv_data.dto
+
+data class AuditLogDto(
+    val id: String,
+    val action: String,
+    val auditType: String,
+    val timestamp: String,
+    val entityId: String,
+    val userId: String,
+)
+
+
