+++ src/main/kotlin/data/csv_data/dto/ProjectDto.kt
@@ -1,0 +1,6 @@
+package data.csv_data.dto
+
+data class ProjectDto(
+    val id: String,
+    val name: String,
+)
