+++ src/main/kotlin/data/csv_data/dto/TaskDto.kt
@@ -1,0 +1,9 @@
+package data.csv_data.dto
+
+data class TaskDto(
+    val id: String,
+    val name: String,
+    val description: String,
+    val projectId: String,
+    val stateId: String,
+)
