+++ src/main/kotlin/data/csv_data/dto/TaskStateDto.kt
@@ -1,0 +1,7 @@
+package data.csv_data.dto
+
+data class TaskStateDto(
+    val id: String,
+    val name: String,
+    val projectId: String,
+)
