+++ src/main/kotlin/data/csv_data/dto/UserDto.kt
@@ -1,0 +1,9 @@
+package data.csv_data.dto
+
+data class UserDto(
+    val id: String,
+    val name: String,
+    val hashedPassword: String,
+    val role: String,
+    val projectIds: List<String>
+)
