--- src/main/kotlin/data/csv_parser/CsvParser.kt
@@ -1,8 +1,0 @@
-package data.csv_parser
-
-typealias CsvData = String
-
-interface CsvParser<T> {
-    fun parse(data: List<CsvData>): List<T>
-    fun serialize(data: List<T>): List<CsvData>
-}
