--- src/main/kotlin/data/csv_parser/TaskStateCsvParser.kt
@@ -1,23 +1,0 @@
-package data.csv_parser
-
-import data.dto.TaskStateDto
-import data.util.TaskStateIndex
-
-class TaskStateCsvParser : CsvParser<TaskStateDto> {
-    override fun parse(data: List<CsvData>): List<TaskStateDto> {
-        return data.drop(1).map { line ->
-            val it = line.split(",")
-            TaskStateDto(
-                id = it[TaskStateIndex.ID],
-                name = it[TaskStateIndex.TITLE],
-                projectId = it[TaskStateIndex.PROJECT_ID],
-            )
-        }
-    }
-
-    override fun serialize(data: List<TaskStateDto>): List<String> {
-        return listOf("id,title,projectId") + data.map { datum ->
-            "${datum.id},${datum.name},${datum.projectId}"
-        }
-    }
-}
