--- src/main/kotlin/data/datasource/DataSource.kt
@@ -1,6 +1,0 @@
-package data.datasource
-
-interface DataSource<T> {
-    fun fetch(): List<T>
-    fun save(data: List<T>)
-}
