--- src/main/kotlin/data/dto/AuditLogDto.kt
@@ -1,12 +1,0 @@
-package data.dto
-
-data class AuditLogDto(
-    val id: String,
-    val action: String,
-    val auditType: String,
-    val timestamp: String,
-    val entityId: String,
-    val userId: String,
-)
-
-
