--- src/main/kotlin/data/dto/ProjectDto.kt
@@ -1,8 +1,0 @@
-package data.dto
-
-import java.util.*
-
-data class ProjectDto(
-    val id: String,
-    val name: String,
-)
