--- src/main/kotlin/data/dto/TaskStateDto.kt
@@ -1,7 +1,0 @@
-package data.dto
-
-data class TaskStateDto(
-    val id: String,
-    val name: String,
-    val projectId: String,
-)
